+++ tests/ConsoleKit/ConsoleIOTest.php
@@ -12,13 +12,14 @@ namespace Tests\ConsoleHelpers\ConsoleKit;
 
 
 use ConsoleHelpers\ConsoleKit\ConsoleIO;
+use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\ChoiceQuestion;
 use Symfony\Component\Console\Question\ConfirmationQuestion;
 
-class ConsoleIOTest extends \PHPUnit_Framework_TestCase
+class ConsoleIOTest extends TestCase
 {
 
 	/**